+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## Unreleased
-### ...
+### Added
+- Output directory parameter to comparison script.
+
+### Fixed
+- Comparison script displays error message when bed file is required but not provided.